Assign custom texture to game cartridge and prune main menu
- Updated JPCartdridge.bin to use SonicCD.png texture across SegaSplash, TitleScreen, and StartGameButton.
- Fixed dark models in menus by switching RenderMesh from MESH_NORMALS to MESH_COLORS.
- Simplified main menu by removing Achievements and Leaderboards options.
- Cleaned up redundant rendering calls in TitleScreen.cpp.

+++ RSDKv4/NativeObjects/SegaSplash.cpp
@@ -169,7 +169,7 @@ void SegaSplash_Main(void *objPtr)
             RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0xFF, 0xFF, 0xFF, 0xFF);
             SetRenderBlendMode(RENDER_BLEND_ALPHA);
             RenderImage(0.0, 0.0, 160.0, 0.4, 0.4, 256.0, 128.0, 512.0, 256.0, 0.0, 0.0, 255, self->textureID);
-            RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0, 0, 0, (int)self->rectAlpha);
+            RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0, 0, 0, self->rectAlpha);
             break;
 
         case SEGAPLASH_STATE_EXIT:
@@ -180,7 +180,7 @@ void SegaSplash_Main(void *objPtr)
             RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0xFF, 0xFF, 0xFF, 0xFF);
             SetRenderBlendMode(RENDER_BLEND_ALPHA);
             RenderImage(0.0, 0.0, 160.0, 0.4, 0.4, 256.0, 128.0, 512.0, 256.0, 0.0, 0.0, 255, self->textureID);
-            RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0, 0, 0, (int)self->rectAlpha);
+            RenderRect(-SCREEN_CENTERX_F, SCREEN_CENTERY_F, 160.0, SCREEN_XSIZE_F, SCREEN_YSIZE_F, 0, 0, 0, self->rectAlpha);
             break;
 
         case SEGAPLASH_STATE_SPAWNCWSPLASH: ResetNativeObject(self, CWSplash_Create, CWSplash_Main); break;
